+++ lib/substation.rb
@@ -36,5 +36,6 @@ end
 require 'substation/request'
 require 'substation/response'
 require 'substation/observer'
+require 'substation/chain'
 require 'substation/dispatcher'
 require 'substation/support/utils'